Fix test_update test
It looks like in 829de290 Updater.updates_cached changed from returning
the length of _cache.files() to just that list directly. The
corresponding test was not updated though because it looks like it will
be skipped unless it's being run as root, which is not the common
configuration.

My local github runner however runs it as root and was failing on this
test.