Remove join() from Service.__del__()
There are no guarantees which thread __del__() is run in. It could
potentially be run in the Service's thread, where it attempts to
join() itself and RuntimeErrors, as captured by this exception I
witnessed during testing:

Exception ignored in: <function Service.__del__ at 0x719dbf61e8c0>
Traceback (most recent call last):
  File "oresat-olaf/olaf/common/service.py", line 48, in __del__
    self._thread.join()
  File "/usr/lib/python3.10/threading.py", line 1093, in join
    raise RuntimeError("cannot join current thread")
RuntimeError: cannot join current thread

Furthermore it doesn't really make sense to wait for the thread to have
stopped in __del__(). We don't do anything with that information and it
doesn't make sense to pause GC until the thread is stopped. Setting the
_event seems enough, the thread can die at its leisure.