A first pass on tests for the Service class
This takes the Service class through its lifecycle, starting, stopping,
and causing errors. Some questions remain about the intended behavior.

In most tests it derives a new class that overrides the specific
behavior that the test requires, but black mangles the formatting of
that so I turn it off for the class creation.