Give ServiceState FAILING and STOPPING different values
This fixes confusion between the two states: when printing it would
always show STOPPING even explicitly set to FAILING, and when checking
for FAILING it would be true when set to just STOPPING.

This additionally changes the type from IntEnum to just Enum. IntEnums
are intended for when the members should also be used interchangeably
with normal ints. Arithmetic, ordering comparisons, and the like. It
doesn't really make sense to use the int properties for ServiceState
except as a way to distinguish them so we shouldn't advertise the option
that it could be done.

Finally @unique is a fun decorator to prevent the kind of error that
occurred here. It ensures that each member has a unique value.