Fix tpdo crash when no cards are on CAN network
This is part of the first in-flight patch to the satellite. We
discovered that when there are no cards on the CAN network this tpdo
will throw an exception. In the patch all cards except for the C3 were
turned off for powersaving reasons.

This clears the `raise_error` flag, and while best practices would have
me write something like `self.send_tpdo(i + 1, raise_error=False)` or a
try-except one of the requirements of the patch was to use the least
amount of bytes possible and this was the most compact way I could
silence the error.