Replace OreSatId with Consts from oresat-configs
A while back I deprecated OreSatId in favor of Consts, which provides a
more feature rich version of OreSatId. This uses those features here in
action, `from_string()` replaces argument handling for naming missions,
`default()` ensures tests are run with the current mission, and
`__str__()` provides a standard human readable name for the rest api.

Getting the Mission via id from the od was harder than it needs to be,
in the next oresat-configs update will introduce a `Consts.from_id()`
similar to `Consts.from_string()` to make this easier.