Handle more eeprom errors
While trying to run OLAF on more people's laptops we ran into another
type of eeprom read error, where it could read the eeprom but the
contents were not a string. This handles that specific case but really
we shouldn't be trying to read arbitrary eeproms if hardware is being
mocked.

I could add a specific eeprom value to look for but that's not yet
common across other projects so this turns it off if any hardware is
mocked.